requirements.txt
absl-py==2.1.0
arrow==1.3.0
asgiref==3.8.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
binaryornot==0.4.4
cached-property==1.5.2
cachetools==5.3.2
caer==2.0.8
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==7.1.2
colorama==0.4.6
cookiecutter==2.4.0
cryptography==42.0.5
dj-database-url==2.1.0
Django==4.2.11
django-browser-reload==1.12.1
django-filter==24.2
django-redis==5.4.0
django-rest-knox==4.2.0
django-tailwind==3.6.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dynaconf==3.2.5
easyocr==1.7.1
exceptiongroup==1.1.3
filelock==3.13.1
flatbuffers==23.5.26
fsspec==2024.2.0
gast==0.5.4
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.0
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
hvac==2.1.0
idna==3.4
imageio==2.33.1
importlib-metadata==7.0.1
Jinja2==3.1.2
keras==2.15.0
knox==0.1.14
lazy_loader==0.3
libclang==16.0.6
loguru==0.7.2
lxml==4.9.4
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
mypy==1.8.0
mypy-extensions==1.0.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.3
oauthlib==3.2.2
opencv-contrib-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
outcome==1.3.0.post0
packages==0.1.1
packaging==24.0
pillow==10.3.0
protobuf==4.23.4
psycopg==3.1.18
psycopg2==2.9.1
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pycparser==2.22
Pygments==2.17.1
PyJWT==2.8.0
pymongo==3.13.0
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
redis==3.5.3
requests==2.31.0
requests-cache==0.5.2
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
scikit-image==0.22.0
scipy==1.12.0
selenium==4.14.0
shapely==2.0.2
shortuuid==1.0.13
six==1.16.0
SMTPEmail==0.4.2
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
text-unidecode==1.3
tifffile==2024.1.30
tomli==2.0.1
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.2
trio==0.22.2
trio-websocket==0.11.1
types-python-dateutil==2.8.19.14
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.0.7
validators==0.28.0
Werkzeug==3.0.1
win32-setctime==1.1.0
wrapt==1.14.1
wsproto==1.2.0
zipp==3.17.0
